requirements.txt
gradio
openai
opencv-python
python-dotenv
requests
Pillow
numpy